Revert "#154"
This reverts commit 0c930531ee402ce081d221d8763973113eeaa1ef.

+++ src/main/java/ru/innopolis/stc12/booksharing/controller/BookEditionsController.java
@@ -71,7 +71,7 @@ public class BookEditionsController {
     }
 
     @GetMapping(value = "/bookEditionDesc/{id}")
-    public String showBookEditionDescriptionPage(@PathVariable long id, Model model) {
+    public String showBookEditionDescriptionPage(@PathVariable int id, Model model) {
         BookEdition bookEdition = bookEditionsService.getById(id);
 
         int countBookCopy = bookCopiesService.getBookCopyCountByBookEditionId(id);
