+++ src/test/java/ru/innopolis/stc12/booksharing/model/dao/BookCopiesDaoImplTest.java
@@ -53,9 +53,9 @@ class BookCopiesDaoImplTest {
         when(jdbcTemplate.update(anyString(), any(Object.class))).thenReturn(0);
         when(bookCopy.getBookEdition()).thenReturn(bookEdition);
 //        when(bookCopy.getOwner()).thenReturn(user);
-        when(bookEdition.getId()).thenReturn(1L);
-        when(user.getId()).thenReturn(1L);
-        when(bookCopy.getBookEdition().getId()).thenReturn(1L);
+        when(bookEdition.getId()).thenReturn(1);
+        when(user.getId()).thenReturn((int) 1);
+        when(bookCopy.getBookEdition().getId()).thenReturn(1);
 //        when(bookCopy.getOwner().getId()).thenReturn((int) 1);
         when(bookCopy.getStatus()).thenReturn(BookCopiesStatus.FREE);
 //todo        bookCopiesDao.save(bookCopy);
